Choose the brand by ranking them
This settles the choice when there are mutiple competing
important brands to use. It implements an algorithm to rank the
brand annotations depending on the predicate used
(isClassifiedBy or implicityClassifiedBy) and a hard-coded priority
list.

Here's an example of when this is particularly useful. An item of
content is annotated with the following:
  * isClassifiedBy Brand "Big Read"
  * isImplicitlyClassifiedBy Brand "Moral Money"
  * isImplicitlyClassifiedBy Brand "Climate Capital"