Add metadata to response payload for tracking
This captures metadata about the content returned in in the
payload, the flag state and the content selection strategies used.

This will be used by the client to send contextual data to spoor,
allowing more accurate analysis of MVT tests.

The code isn't very clean at the moment. We can return to that as
part of some refatoring we need to do.