Change onwardJourneyTests=variant2 behaviour
Variant2 will now act a a kind of control to Variant 1.

The control introduces no changes to the content selected by the API.

Variant 1 changes the way content is selected and in some
cases it also results in more recommendations being displayed to
the reader.

Previously Variant 2 was a mild difference to Variant1 (described
above).

With this change, Variant 2 will now select content in almost
the same way to the control but display more stories at the same rate
as Variant1. This helps us know if simply showing more stories
caused a difference.