Remove Ribbon slot
The component has already been removed from the front
end (see (`next-article` #4575)[https://github.com/Financial-Times/next-article/pull/4575]).

The ribbon was a special case and not like the other two slots.
For example, it duplicates were not removed from the Ribbon.
So we may have to bring back some of this logic in the future
if we invent new slots.

For now, removing this code allows us to refactor the app so
it's cleaner, easier to test and extend.